broke InterpreterTests into multiple files
